Single Lane Bridge Problem
Single-lane bridge connects the two Vermont villages of North Tun- Bridge and South Tunbridge. Farmers in the two villages use this bridge to deliver their produce to the neighboring town.  The bridge can become deadlocked if a northbound and a southbound farmer get on the bridge at the same time. (Vermont farmers are stubborn and are unable to back up.) Using semaphores and/or mutex locks, design an algorithm in pseudocode that prevents deadlock. Initially, do not be concerned about starvation (the situation in which northbound farmers prevent southbound farmers from using the bridge, or vice versa).
Modify your solution, so that it is starvation-free.